+++ engines/tahi_standard_tasks/app/models/tahi_standard_tasks/tech_check_task.rb
@@ -1,9 +1,0 @@
-module TahiStandardTasks
-  class TechCheckTask < Task
-    register_task default_title: 'Tech Check', default_role: 'admin'
-
-    def active_model_serializer
-      TaskSerializer
-    end
-  end
-end