+++ engines/tahi_standard_tasks/client/app/views/overlays/tech-check.js
@@ -1,6 +1,0 @@
-import OverlayView from 'tahi/views/overlay';
-
-export default OverlayView.extend({
-  templateName: 'overlays/tech-check',
-  layoutName: 'layouts/overlay'
-});