+++ spec/factories/decision_factory.rb
@@ -3,6 +3,6 @@ FactoryGirl.define do
     association :paper, factory: :paper
     sequence(:revision_number) { |n| n }
     letter 'Test Decision Letter'
-    verdict 'Test verdict'
+    verdict 'accept'
   end
 end